backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
pydantic==2.0.3
tortoise-orm==0.20.0
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.9.1
isort==5.12.0
